requirements.txt
streamlit
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
langchain
langchain_community
langchain-groq